Use BIGINT-compatible ChoiceCode in allocation models

+++ api/Models/Allocation/AllocationModels.cs
@@ -58,7 +58,7 @@ public sealed class CollegePreference
 {
     public int PreferenceNo { get; set; }
     public int CollegeId { get; set; }
-    public int ChoiceCode { get; set; }
+    public long ChoiceCode { get; set; }
 }
 
 public sealed class ExistingAllotment
@@ -69,13 +69,13 @@ public sealed class ExistingAllotment
     public string AllocatedType { get; set; } = string.Empty;
     public string OriginalAllocatedType { get; set; } = string.Empty;
     public string VacancyType { get; set; } = string.Empty;
-    public int ChoiceCode { get; set; }
+    public long ChoiceCode { get; set; }
 }
 
 public sealed class SeatInventory
 {
     public int CollegeId { get; set; }
-    public int ChoiceCode { get; set; }
+    public long ChoiceCode { get; set; }
     public int CategoryId { get; set; }
     public int QuotaId { get; set; }
     public int General { get; set; }
@@ -96,7 +96,7 @@ public sealed class AllocationDecision
     public string AllocatedType { get; set; } = string.Empty;
     public string OriginalAllocatedType { get; set; } = string.Empty;
     public string VacancyType { get; set; } = string.Empty;
-    public int ChoiceCode { get; set; }
+    public long ChoiceCode { get; set; }
     public int StepId { get; set; }
     public string DecisionArea { get; set; } = string.Empty;
     public string RuleCode { get; set; } = string.Empty;
